tech(importCsv): feedbacks PR
