+++ api/controllers/CaveController.js
@@ -126,7 +126,6 @@ module.exports = {
       }
     };
 
-    // Launch creation request using transaction: it performs a rollback if an error occurs
     const caveCreated = await CaveService.createCave(
       cleanedData,
       nameAndDescData,
