--- api/controllers/v1/ImportCsvController.js
@@ -1,6 +1,0 @@
-const importCsvController = require('../ImportCsvController');
-
-module.exports = {
-  checkAll: (req, res) => importCsvController.checkAll(req, res),
-  importAll: (req, res) => importCsvController.importAll(req, res),
-};
