+++ assets/js/react/components/appli/ImportCSV/constants.js
@@ -1,0 +1,2 @@
+export const ENTRANCE = 0;
+export const DOCUMENT = 1;
