+++ assets/js/react/components/appli/ImportCSV/types.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 
 const defaultFormSteps = [
   { id: 1, name: 'General Information', isValid: false },
-  { id: 2, name: 'File completion', isValid: false },
+  { id: 2, name: 'File Completion', isValid: false },
   { id: 3, name: 'Review', isValid: false },
   { id: 4, name: 'Confirm or Rollback', isValid: true },
 ];
