+++ assets/js/react/pages/ImportCSV.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ImportContainer from '../components/common/ImportCSV/ImportContainer';
+import ImportContainer from '../components/appli/ImportCSV/ImportContainer';
 
 const ImportCSV = () => <ImportContainer />;
