+++ package.json
@@ -59,6 +59,7 @@
     "react-intl": "^5.10.16",
     "react-is": "^16.12.0",
     "react-leaflet": "^3.1.0",
+    "react-papaparse": "^3.17.1",
     "react-redux": "^7.2.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
